#2 Slimifying BE.
