+++ backend/package.json
@@ -19,6 +19,7 @@
     "@ty-ras/metadata-jsonschema-io-ts": "0.12.1",
     "@ty-ras/metadata-openapi": "0.12.2",
     "@ty-ras/resource-pool-fp-ts": "0.12.0",
+    "@ty-ras/typed-sql-io-ts": "0.12.0",
     "aws-jwt-verify": "3.1.0",
     "fp-ts": "2.13.1",
     "io-ts": "2.2.19",
