--- backend/src/services/db/index.ts
@@ -1,3 +1,0 @@
-// TODO move this folder to generic library.
-export * from "./query";
-export * from "./validation";
